% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOJ.R
\name{get_boj_datasets}
\alias{get_boj_datasets}
\title{Download data a frame listing available BOJ data sets}
\usage{
get_boj_datasets()
}
\value{
A tibble data frame
}
\description{
Download data a frame listing available BOJ data sets
}
\examples{
datasets <- get_boj_datasets()
}
