% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOJ.R
\name{get_boj}
\alias{get_boj}
\title{Download and parse a BOJ data set}
\usage{
get_boj(url, ...)
}
\arguments{
\item{url}{URL of the data set to be imported (usually obtained through
\code{get_boj_datasets()})}

\item{...}{Arguments passed to \code{download.file()} (e.g.
\code{quiet = TRUE})}
}
\value{
A tibble data frame
}
\description{
Download and parse a BOJ data set
}
\examples{
\dontrun{
datasets <- get_boj_datasets()
df <- get_boj(datasets$url[(datasets$name == "sppi_q_en")])
}
}
