% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOJ.R
\name{get_boj_datasets}
\alias{get_boj_datasets}
\title{Download and parse a list of available BOJ data sets}
\usage{
get_boj_datasets(
  base_url = "https://www.stat-search.boj.or.jp/info/dload_en.html"
)
}
\arguments{
\item{base_url}{Character. URL of the BOJ's Time-Series Data portal flat
files page (optional).}
}
\value{
A tibble data frame.
}
\description{
Download and parse a list of available BOJ data sets
}
\examples{
\donttest{
ds <- get_boj_datasets()
}
}
